Revert "cmass/bias/"
This reverts commit 619387347a80517b9adb7bb25e6c55fd88c27e04.
