+++ cmass/bias/tools/hod.py
@@ -12,10 +12,11 @@ which additionally uses the `Hod_parameter` helper class
 for each parameter.
 """
 
+from halotools.custom_exceptions import HalotoolsError
+from scipy.special import erf
 import logging
 import numpy as np
 from omegaconf import open_dict
-from .hod_models import Zheng07zdepCens, Zheng07zdepSats
 
 from halotools.empirical_models import NFWProfile
 from halotools.sim_manager import UserSuppliedHaloCatalog
@@ -479,6 +480,38 @@ def build_halo_catalog(
     return UserSuppliedHaloCatalog(**kws)
 
 
+def logM_i(z, logM_i_pivot, mu_i_p, z_pivot):
+    return logM_i_pivot + mu_i_p * ((1 / (1 + z)) - (1 / (1 + z_pivot)))
+
+
+class Zheng07zdepCens(Zheng07Cens):
+    # Params: logMmin, sigma_logM, mumin, zpivot
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+        self.param_dict['mumin'] = 0.0
+        self.param_dict['zpivot'] = 0.5
+
+        self.list_of_haloprops_needed = ['halo_redshift']
+
+    def mean_occupation(self, **kwargs):
+        # Retrieve the array storing the mass-like variable
+        mass = kwargs["table"][self.prim_haloprop_key]
+        redshift = kwargs["table"]["halo_redshift"]
+
+        logM = np.log10(mass)
+        logMmin = logM_i(
+            redshift, self.param_dict["logMmin"], self.param_dict["mumin"],
+            self.param_dict["zpivot"]
+        )
+        mean_ncen = 0.5 * (
+            1.0
+            + erf((logM - logMmin) / self.param_dict["sigma_logM"])
+        )
+
+        return mean_ncen
+
+
 def build_HOD_model(
     cosmology, model, theta, zf, mdef='vir',
 ):
@@ -515,7 +548,7 @@ def build_HOD_model(
         )
     elif model == 'zheng07zdep':
         cenocc = Zheng07zdepCens(prim_haloprop_key=mkey)
-        satocc = Zheng07zdepSats(
+        satocc = Zheng07Sats(
             prim_haloprop_key=mkey,
             cenocc_model=cenocc,
             modulate_with_cenocc=True
